backend/requirements.txt
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.7.0
requests==2.25.1
python-dotenv==0.19.0
pg8000==1.21.1
numpy==1.23.5
pandas==1.4.3
scikit-learn==1.1.3
joblib==1.2.0
matplotlib==3.5.2
seaborn==0.11.2
SQLAlchemy==1.4.0